+++ src/main/java/com/connection/emobile/repository/MobileNumberRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.connection.emobile.entity.MobileNumber;
 
 @Repository
-public interface MobileNumberRepository extends JpaRepository<MobileNumber, Integer> {
+public interface MobileNumberRepository extends JpaRepository<MobileNumber, Long> {
 
 }
